# perf: Perform Local Distinct list_agg in count_distinct aggregations